Fix node corruption bug while cleaning uploads
When we clean a node, we ask nginx to free the memory. On some
platform or no load there were great chance the memory wasn't
reused for something else.
With high load and on FreeBSD, the memory seems to be reused
immediatly and we jump in the middle of f*****g nowhere, and
we crash.

Thanks to Vladimir Getmanshchuk for his bug report.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>